ref(utils): Move number formatters into their own files
