+++ static/app/components/events/eventTagsAndScreenshot/screenshot/modal.tsx
@@ -16,7 +16,8 @@ import {t, tct} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import type {EventAttachment, IssueAttachment, Organization, Project} from 'sentry/types';
 import type {Event} from 'sentry/types/event';
-import {defined, formatBytesBase2} from 'sentry/utils';
+import {defined} from 'sentry/utils';
+import {formatBytesBase2} from 'sentry/utils/bytes/formatBytesBase2';
 import getDynamicText from 'sentry/utils/getDynamicText';
 import parseLinkHeader from 'sentry/utils/parseLinkHeader';
 import useApi from 'sentry/utils/useApi';
