+++ static/app/utils/metrics/formatters.tsx
@@ -1,10 +1,11 @@
 import {t} from 'sentry/locale';
 import type {MetricType} from 'sentry/types/metrics';
-import {defined, formatBytesBase2, formatBytesBase10} from 'sentry/utils';
+import {defined} from 'sentry/utils';
+import {formatBytesBase2} from 'sentry/utils/bytes/formatBytesBase2';
+import {formatBytesBase10} from 'sentry/utils/bytes/formatBytesBase10';
 import {
   DAY,
   formatAbbreviatedNumberWithDynamicPrecision,
-  formatNumberWithDynamicDecimalPoints,
   HOUR,
   MICROSECOND,
   MILLISECOND,
@@ -14,6 +15,7 @@ import {
   SECOND,
   WEEK,
 } from 'sentry/utils/formatters';
+import {formatNumberWithDynamicDecimalPoints} from 'sentry/utils/number/formatNumberWithDynamicDecimalPoints';
 
 const metricTypeToReadable: Record<MetricType, string> = {
   c: t('counter'),
