+++ static/app/views/performance/newTraceDetails/traceDrawer/details/transaction/sections/cacheMetrics.tsx
@@ -1,5 +1,5 @@
 import {t} from 'sentry/locale';
-import {formatBytesBase2} from 'sentry/utils';
+import {formatBytesBase2} from 'sentry/utils/bytes/formatBytesBase2';
 import {formatPercentage} from 'sentry/utils/formatters';
 import type {SpanMetricsResponse} from 'sentry/views/starfish/types';
 import {DataTitles} from 'sentry/views/starfish/views/spans/types';
