+++ static/app/views/replays/detail/memoryPanel/memoryChart.tsx
@@ -11,8 +11,8 @@ import type {useReplayContext} from 'sentry/components/replays/replayContext';
 import {formatTime} from 'sentry/components/replays/utils';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import {formatBytesBase2} from 'sentry/utils';
 import toArray from 'sentry/utils/array/toArray';
+import {formatBytesBase2} from 'sentry/utils/bytes/formatBytesBase2';
 import {getFormattedDate} from 'sentry/utils/dates';
 import domId from 'sentry/utils/domId';
 import type {MemoryFrame} from 'sentry/utils/replays/types';
